requirements.txt
torch
numpy
transformers
accelerate
liger-kernel-nightly